+++ NEWS.md
@@ -13,6 +13,10 @@ arg News and Updates
 
 * Fixed a bug where the message provided to `.msg` (if any) was not evaluated in the correct environment.
 
+* Fixed several typos in the documentation and in some error messages.
+
+* Added a testing suite.
+
 # *arg* 0.1.1
 
 * `err()`, `wrn()`, and `msg()` now pass arguments specified in `...` to `rlang::abort()`, `rlang::warn()`, and `rlang::inform()`, respectively.