+++ NEWS.md
@@ -15,6 +15,8 @@ arg News and Updates
 
 * Fixed several typos in the documentation and in some error messages.
 
+* Added a new logo.
+
 * Added a testing suite.
 
 # *arg* 0.1.1